requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
langchain>=0.2.7
langchain-community>=0.2.7
langchain-openai>=0.1.16
langchain-text-splitters>=0.2.0
chromadb>=0.5.5
pypdf>=4.2.0
docx2txt>=0.8
beautifulsoup4>=4.12.3
lxml>=5.2.2
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
python-multipart>=0.0.9
pydantic>=2.7.0
